tweak: dinheiro deve estar em int
Se for preciso de casas decimais, então medimos centavos de spesos

+++ Content.Client/_Gabystation/CartridgeLoader/Cartridges/NanoBankUi.cs
@@ -33,7 +33,7 @@ public sealed partial class NanoBankUi : UIFragment
 
     private static void SendNanoBankUiMessage(NanoBankUiMessageType type,
         int? targetAcc,
-        float? content,
+        int? content,
         BoundUserInterface userInterface)
     {
         var nanoChatMessage = new NanoBankUiMessageEvent(type, targetAcc, content);
